+++ modules/book.js
@@ -1,8 +1,6 @@
-class Books {
+export default class Books {
   constructor(title, author) {
     this.title = title;
     this.author = author;
   }
 }
-
-export default Books;
